Half-viability period

P50 computes the half-viability period, which is the time taken for 50% of the seeds to lose viability.

P50(initial, vcindex, vcdirect, mc, temp, years = FALSE)

Arguments

initial
The initial viability (%).

vcindex
An integer value indicating the index of seed viability. constants to be used from the viabilityconstants dataset in the package.

vcdirect
A numeric vector of length 4 with the four viability constants viz.: KE, CW, CH and CQ.

mc
Moisture content.

temp
Temperature in °C.

years
If TRUE, returns the output period in years instead of days.

Value

The half-viability period in days or years (according to argument years).

Details

The period to lose 50% viability (P50) is computed according to the relationship between probit percentage viabilities and time of storage described by Ellis and Roberts (1980) as follows.

v = Ki − [ p ⁄ σ ]

or

v = Ki − (1 ⁄ σ)⋅p

Where, v is the probit percentage viability at storage time p (final viability), Ki is the probit percentage viability of the seedlot at the beginning of storage (seedlot constant) and 1⁄σ is the slope.

Germination percentages plotted against storage times yield a sigmoid seed survival curve which is converted to a linear relationship by the probit transformation with slope 1 ⁄ σ.

When v = 0 (equivalent to 50% viability), P50 can be computed as follows.

P50 = Ki × σ

If the initial viablity (initial) is beyond limits (0-100 %, an error is issued.

The value of this computation is appropriate for temperature between -20 to 90 °C and seed moisture content between 5 to 25%. For values beyond this range, a warning will be displayed.

Note

For initial viability percentage values of 100%, adjust it according to sample size using the PercentAdjust function to avoid infinity values in output.

References

Ellis RH, Roberts EH (1980). “Improved equations for the prediction of seed longevity.” Annals of Botany, 45(1), 13--30.

See also

Examples

P50(initial = 98, vcindex = 24, mc = 5, temp = -20)

#> [1] 1960829

P50(initial = 98, vcindex = 24, mc = 5, temp = -20, years = TRUE)

#> [1] 5372.134

# With intial viability of 100%
P50(initial = 100, vcindex = 24, mc = 5, temp = -20)

#> [1] Inf

P50(initial = 100, vcindex = 24, mc = 5, temp = -20, years = TRUE)

#> [1] Inf

# With intial viability of 100%, use of PercentAdjust() to avoid Inf
P50(initial = PercentAdjust(100, n = 50), vcindex = 24, mc = 5, temp = -20)

#> [1] 2459288

P50(initial = PercentAdjust(100, n = 50), vcindex = 24, mc = 5, temp = -20,
    years = TRUE)

#> [1] 6737.777

# \dontrun{
# Error if initial viability is beyond limits (0-100 %)
P50(initial = 110, vcindex = 24, mc = 5, temp = -20)

#> Error in P50(initial = 110, vcindex = 24, mc = 5, temp = -20): "initial" is beyond limits (0 < "initial" < 100).

# }