Seed equilibrium moisture content and equilibrium relative humidity of the seed storage environment

Compute the following metrics:

EqRH

The equilibrium relative humidity of the seed storage environment from seed equilibrium moisture content, oil content and temperature.

SeedEqMC(oilcontent, rh, temp, basis = c("wet", "dry"))

EqRH(oilcontent, mc, temp, basis = c("wet", "dry"))

Arguments

oilcontent
The percentage oil content of seed (dry basis).

rh
Relative humidity expressed in percentage.

temp
Temperature in °C.

basis
The type of estimation of moisture content specified in the argument mc. Either "wet" or "dry".

mc
The seed equilibrium moisture content on wet or dry basis (according to argument basis).

Value

For SeedEqMC, the seed equilibrium moisture content on wet or dry basis (according to argument basis) expressed in percentage.

For EqRH, the equilibrium relative humidity expressed in percentage.

Details

This relationship between seed equilibrium moisture content, seed oil content, the equilibrium relative humidity and temperature of the storage environment was described by Cromarty et al. (1982) as follows.

(1-R) = e −[ [ (Me × (1.1 + T⁄90)) ⁄ (1 − Do) ]2 ⁄ 440 ]

Where, R is the relative humidity expressed as decimal, Me is the equilibrium percentage moisture content (dry basis), T is the temperature in °C of air or the seed equilibrium, DO is the oil content of seed (dry basis) expressed as decimal and e is the mathematical constant 2.718282.

For values of oil content (oilcontent), relative humidity (rh) and seed equilibrium moisture content (mc) beyond the limits of 0-100 %, a warning is issued.

Note

The above expression by Cromarty et al. (1982) is recommended for temperature and humidity ranges of 0-40 °C and 10-70% RH for starchy seeds (eg. cereals); and 15-25 °C and 10-70% RH for oilseeds.

References

Cromarty AS, Ellis RH, Roberts EH (1982). The Design of Seed Storage Facilities for Genetic Conservation, Revised 1985 and 1990 edition. International Board for Plant Genetic Resources, Rome, Italy.

See also

Examples

SeedEqMC(oilcontent = 29, rh = 13, temp = 25, basis = "wet")

#> [1] 3.877459

SeedEqMC(oilcontent = 29, rh = 13, temp = 25, basis = "dry")

#> [1] 4.03387

EqRH(oilcontent = 29, mc = 5, temp = 25, basis = "wet")

#> [1] 21.10662

EqRH(oilcontent = 29, mc = 5, temp = 25, basis = "dry")

#> [1] 19.26179

# Warning if oilcontent is beyond limits (0-100 %)
SeedEqMC(oilcontent = 125, rh = 13, temp = 25, basis = "wet")

#> Warning: "oilcontent" is beyond limits (0 < "oilcontent" < 100)

#> Warning: "mc" is beyond limits (0 < "mc" < 100)

#> [1] -1.440842

EqRH(oilcontent = 125, mc = 5, temp = 25, basis = "wet")

#> Warning: "oilcontent" is beyond limits (0 < "oilcontent" < 100)

#> [1] 85.22354

# Warning if relative humidity is beyond limits (0-100 %)
SeedEqMC(oilcontent = 29, rh = 115, temp = 25, basis = "wet")

#> Warning: "rh" is beyond limits (0 < "rh" < 100)

#> Warning: NaNs produced

#> [1] NaN

# Warning if moisture content is beyond limits (0-100 %)
EqRH(oilcontent = 29, mc = 115, temp = 25, basis = "wet")

#> Warning: "mc" is beyond limits (0 < "mc" < 100)

#> Warning: "mc" is beyond limits (0 < "mc" < 50)

#> [1] 100