Period to lose unit probit viability

Sigma calculates the period to lose one probit viability (σ) under storage at a given moisture content and temperature.

Sigma(vcindex, vcdirect, mc, temp, years = FALSE)

Arguments

vcindex
An interger value indicating the index of seed viability. constants to be used from the viabilityconstants dataset in the package.

vcdirect
A numeric vector of length 4 with the four viability constants viz.: KE, CW, CH and CQ.

mc
Moisture content.

temp
Temperature in °C.

years
If TRUE, returns the output period in years instead of days.

Value

The period to lose one probit in days or years (according to argument years).

Details

This function computes the period to lose one probit viability (σ) according to the improved seed viability equation of Ellis and Roberts (1980) as follows:

v = Ki − [ p ⁄ σ ]

or

v = Ki − (1 ⁄ σ)⋅p

Where, v is the probit percentage viability at storage time p (final viability), Ki is the probit percentage viability of the seedlot at the beginning of storage (seedlot constant) and 1⁄σ is the slope.

Germination percentages plotted against storage times yield a sigmoid seed survival curve which is converted to a linear relationship by the probit transformation with slope 1 ⁄ σ.

The slope is determined as follows:

σ = 10 KE − CW logm − CHt − CQt2

Where, v is the probit percentage viability at storage time p (final viability), Ki is the probit percentage viability of the seedlot at the beginning of storage (seedlot constant), m is the moisture content (fresh weight basis), t is the temperature and KE, CW, CH and CQ are the species-specific seed viability constants.

The value of the species-specific seed viability constants can be specified either directly in the arguement vcdirect or as the index value of the required seed viability constants from the viabilityconstants dataset through the argument vcindex.

The value of this prediction is appropriate for temperature between -20 to 90 °C and seed moisture content between 5 to 25%. For values beyond this range, a warning will be displayed.

References

Ellis RH, Roberts EH (1980). “Improved equations for the prediction of seed longevity.” Annals of Botany, 45(1), 13--30.

Examples

#----------------------------------------------------------------------------
# Days/Years to lose unit probit viability for rice seeds stored at
# 5 degree celsius and 10% moisture content.
#----------------------------------------------------------------------------
# Fetch the index from viabilityconstants dataset
viabilityconstants[grepl("oryza", x = viabilityconstants$Species,
                         ignore.case = TRUE),]

#> # A tibble: 8 x 7
#>   Index Species              Ke    Cw     Ch       Cq Reference                
#>   <dbl> <chr>             <dbl> <dbl>  <dbl>    <dbl> <chr>                    
#> 1    85 Oryza glaberrima   9.41  5.04 0.0375 0.000471 Ellis et al., 2007       
#> 2    86 Oryza glaberrima*  6.87  5.51 0.0329 0.000478 Bam, Hong, Ellis, Kumaga~
#> 3    87 Oryza sativa       8.24  4.34 0.0307 0.000501 Ellis et al., 2007       
#> 4    88 Oryza sativa*      8.67  5.03 0.0329 0.000478 Ellis et al., 1989       
#> 5    89 Oryza sativa ssp~  8.81  4.90 0.0329 0.000478 Ellis, Hong, and Roberts~
#> 6    90 Oryza sativa ssp~  8.42  4.90 0.0329 0.000478 Ellis et al., 1992       
#> 7    91 Oryza sativa ssp~  6.63  5.51 0.0329 0.000478 Bam et al., 2008         
#> 8    92 Oryza sativa ssp~  8.74  4.90 0.0329 0.000478 Ellis et al., 1992       

# Use index 87
Sigma(vcindex = 87, mc = 10, temp = 5)

#> [1] 5382.388

Sigma(vcindex = 87, mc = 10, temp = 5, years = TRUE)

#> [1] 14.74627

# Input the viability constants directly
Sigma(vcdirect = c(8.242, 4.345, 0.0307, 0.000501), mc = 10, temp = 5)

#> [1] 5382.388

Sigma(vcdirect = c(8.242, 4.345, 0.0307, 0.000501), mc = 10, temp = 5,
      years = TRUE)

#> [1] 14.74627

# Warning if moisture content is beyond limits (0-100 %)
Sigma(vcindex = 87, mc = 110, temp = 5)

#> Warning: "mc" is beyond limits (0 < "mc" < 100)

#> [1] 0.1607399

# Warning if temperature is beyond limits (-20 to 90 degree C)
Sigma(vcindex = 87, mc = 10, temp = 95)

#> Warning: "temp" is beyond limits (-20 < "temp" < 90)

#> [1] 0.0002877233

#----------------------------------------------------------------------------
# Days/Years to lose unit probit viability for soybean seeds stored at
# -18 degree celsius and 8% moisture content.
#----------------------------------------------------------------------------
# Fetch the index from viabilityconstants dataset
viabilityconstants[grepl("glycine", x = viabilityconstants$Species,
                         ignore.case = TRUE),]

#> # A tibble: 2 x 7
#>   Index Species        Ke    Cw     Ch      Cq Reference                       
#>   <dbl> <chr>       <dbl> <dbl>  <dbl>   <dbl> <chr>                           
#> 1    59 Glycine max  7.52  4.09 0.0329 4.78e-4 Ellis et al., 1982; Dickie et a~
#> 2    60 Glycine max  7.29  4.00 0.0295 4.91e-4 Ellis, 1988                     

# Use index  59
Sigma(vcindex = 59, mc = 8, temp = -18)

#> [1] 18719.85

Sigma(vcindex = 59, mc = 8, temp = -18, years = TRUE)

#> [1] 51.28725

# Input the viability constants directly
Sigma(vcdirect = c(7.292, 3.996, 0.0295, 0.000491), mc = 8, temp = -18)

#> [1] 11354.51

Sigma(vcdirect = c(7.292, 3.996, 0.0295, 0.000491), mc = 8, temp = -18,
      years = TRUE)

#> [1] 31.10825

# Warning if moisture content is beyond limits (0-100 %)
Sigma(vcindex = 59, mc = 110, temp = 5)

#> Warning: "mc" is beyond limits (0 < "mc" < 100)

#> [1] 0.1017231

# Warning if temperature is beyond limits (-20 to 90 degree C)
Sigma(vcindex = 59, mc = 10, temp = 95)

#> Warning: "temp" is beyond limits (-20 < "temp" < 90)

#> [1] 9.989644e-05