SigmaTransformed
Source: R/SigmaTransformed.R

SigmaTransformed transforms the measured sigma value at a specific temperature to an estimate of sigma at another temperature. This useful in comparitive seed testing protocol to compare seed longevities among species tested at different temperatures (Probert et al., 2009).

SigmaTransformed(sigma, temp1, temp2, temp.coeff = c(0.0329, 0.000478))

Arguments

sigma
The inverse of slope from the seed viability equation (σ), estimated at temperature temp1.

temp1
The temperature at which sigma is estimated in °C.

temp2
The temperature at which the transformed σ is to be estimated in °C.

temp.coeff
The species-specific temperature coefficients (CH and CQ.) as a numeric vector of length 2.

Value

The transformed value of σ at temperature temp2.

Details

The transformation is based on the effect of temperature on seed longevity (σ) (identified by storage experiment with constant moisture content and varying temperature) which is as follows:

logσ = β − CHt − CQt2

Where, CH and CQ are the species-specific temperature coefficients, t is the temperature and β is the constant associated with moisture relations of seed longevity.

References

Ellis R and Roberts E (1980). “Improved equations for the prediction of seed longevity.” Annals of Botany, 45(1), pp. 13--30. Probert RJ, Daws MI and Hay FR (2009). “Ecological correlates of ex situ seed longevity: A comparative study on 195 species.” Annals of Botany, 104(1), pp. 57--69. doi: 10.1093/aob/mcp082 , https://academic.oup.com/aob/article/104/1/57/122605.

See also

Sigma

Examples

SigmaTransformed(250, 60, 45)

#> [1] 4408.492