Viability constant KE based on the species-specific temperature coefficients
Source: R/Ke.R

Ke computes the viability constant KE from the species-specific temperature coefficients in case of storage experiments conducted at constant temperature and varying moisture contents.

Ke(K, temp, temp.coeff = c(0.0329, 0.000478))

Arguments

K
The constant K associated with the relationship of temperature with seed longevity (see Details).

temp
Temperature in °C.

temp.coeff
The species-specific temperature coefficients (CH and CQ.) as a numeric vector of length 2.

Value

The value of species-specific seed viability constant KE.

Details

From seed storage experiments involving storage of seeds at a constant temperature in a range of moisture contents, the effect of moisture content on seed longevity (σ) can be estimated from the following linear relationship:

logσ = K − CWlogm

Where, K is the intercept, CW is the slope and m is the moisture content.

The constant K associated with the relationship of temperature with seed longevity as follows:

K = KE − CHt − CQt2

Where, KE, CH and CQ are the species-specific seed viability constants.

The constant KE can be estimated from the universal temperature constants (CH = 0.0329 and CQ = 0.000478) in case of seed storage experiments at constant temperature and varying moisture content as follows:

KE = K + CHt + CQt2

References

Ellis R and Roberts E (1980). “Improved equations for the prediction of seed longevity.” Annals of Botany, 45(1), pp. 13--30. Pritchard HW and Dickie JB (2003). “Predicting seed longevity: The use and abuse of seed viability equations.” In Smith RD, Dickie JB, Linington SH, Pritchard HW and Probert RJ (eds.), Seed Conservation: Turning Science into Practice, pp. 653--721. Kew, UK, Royal Botanic Gardens.

See also

Sigma

Examples

Ke(36, 10)

#> [1] 36.3768