Probit transformation

These functions transform data between percentage, probit and Normal Equivalent Deviate (NED) (Bliss 1934; Finney 1952) .

PercentAdjust(percentage, n)

Percent2NED(percentage)

Percent2Probit(percentage)

Probit2NED(probit)

NED2Probit(NED)

NED2Percent(NED)

Probit2Percent(probit)

Arguments

percentage
The percentage value.

n
Sample size for estimation of percentage.

probit
The probit value

NED
The NED value.

Value

The transformed value.

Details

Probit transformation can be used to transform a sigmoid curve of percentage data to a linear one. The probit transformation is defined as NED + 5. However the two terms probit and NED are used interchangeably in literature.

NED function (Φ-1) is is the inverse of the cumulative distribution function (Φ) of the standard normal distribution (z ∼ N(0,1)) or the quantile function associated with the standard normal distribution.

For percentage p,

NED(p)= Φ-1(p) = √2 erf-1 (2p-1)

and

probit(p) = NED(p) + 5

The PercentAdjust function adjusts the percentage values of 0 and 100 to 100 × [ 0.25 ⁄ n ] and 100 × [ (n − 0.25) ⁄ n ] respectively, according to the sample size n to avoid infinity values during probit transformation (Miller and Tainter 1944) .

References

Bliss CI (1934). “The method of probits.” Science, 79(2037), 38--39.

Miller LC, Tainter ML (1944). “Estimation of the ED50 and its error by means of logarithmic-probit graph paper.” Proceedings of the Society for Experimental Biology and Medicine, 57(2), 261--264.

Finney DJ (1952). Probit Analysis: A Statistical Treatment of the Sigmoid Response Curve. Cambridge University Press, Cambridge, England.

Examples

Percent2NED(0:100)

#>   [1]        -Inf -2.32634787 -2.05374891 -1.88079361 -1.75068607 -1.64485363
#>   [7] -1.55477359 -1.47579103 -1.40507156 -1.34075503 -1.28155157 -1.22652812
#>  [13] -1.17498679 -1.12639113 -1.08031934 -1.03643339 -0.99445788 -0.95416525
#>  [19] -0.91536509 -0.87789630 -0.84162123 -0.80642125 -0.77219321 -0.73884685
#>  [25] -0.70630256 -0.67448975 -0.64334541 -0.61281299 -0.58284151 -0.55338472
#>  [31] -0.52440051 -0.49585035 -0.46769880 -0.43991317 -0.41246313 -0.38532047
#>  [37] -0.35845879 -0.33185335 -0.30548079 -0.27931903 -0.25334710 -0.22754498
#>  [43] -0.20189348 -0.17637416 -0.15096922 -0.12566135 -0.10043372 -0.07526986
#>  [49] -0.05015358 -0.02506891  0.00000000  0.02506891  0.05015358  0.07526986
#>  [55]  0.10043372  0.12566135  0.15096922  0.17637416  0.20189348  0.22754498
#>  [61]  0.25334710  0.27931903  0.30548079  0.33185335  0.35845879  0.38532047
#>  [67]  0.41246313  0.43991317  0.46769880  0.49585035  0.52440051  0.55338472
#>  [73]  0.58284151  0.61281299  0.64334541  0.67448975  0.70630256  0.73884685
#>  [79]  0.77219321  0.80642125  0.84162123  0.87789630  0.91536509  0.95416525
#>  [85]  0.99445788  1.03643339  1.08031934  1.12639113  1.17498679  1.22652812
#>  [91]  1.28155157  1.34075503  1.40507156  1.47579103  1.55477359  1.64485363
#>  [97]  1.75068607  1.88079361  2.05374891  2.32634787         Inf

Percent2Probit(0:100)

#>   [1]     -Inf 2.673652 2.946251 3.119206 3.249314 3.355146 3.445226 3.524209
#>   [9] 3.594928 3.659245 3.718448 3.773472 3.825013 3.873609 3.919681 3.963567
#>  [17] 4.005542 4.045835 4.084635 4.122104 4.158379 4.193579 4.227807 4.261153
#>  [25] 4.293697 4.325510 4.356655 4.387187 4.417158 4.446615 4.475599 4.504150
#>  [33] 4.532301 4.560087 4.587537 4.614680 4.641541 4.668147 4.694519 4.720681
#>  [41] 4.746653 4.772455 4.798107 4.823626 4.849031 4.874339 4.899566 4.924730
#>  [49] 4.949846 4.974931 5.000000 5.025069 5.050154 5.075270 5.100434 5.125661
#>  [57] 5.150969 5.176374 5.201893 5.227545 5.253347 5.279319 5.305481 5.331853
#>  [65] 5.358459 5.385320 5.412463 5.439913 5.467699 5.495850 5.524401 5.553385
#>  [73] 5.582842 5.612813 5.643345 5.674490 5.706303 5.738847 5.772193 5.806421
#>  [81] 5.841621 5.877896 5.915365 5.954165 5.994458 6.036433 6.080319 6.126391
#>  [89] 6.174987 6.226528 6.281552 6.340755 6.405072 6.475791 6.554774 6.644854
#>  [97] 6.750686 6.880794 7.053749 7.326348      Inf

Percent2NED(25)

#> [1] -0.6744898

Percent2Probit(25)

#> [1] 4.32551

Percent2NED(25) +5

#> [1] 4.32551

NED2Probit(-0.6744898)

#> [1] 4.32551

# Percentage adjustment for 0 and 100
Percent2Probit(100)

#> [1] Inf

Percent2Probit(0)

#> [1] -Inf

n = 50
Percent2Probit(PercentAdjust(100, n))

#> [1] 7.575829

Percent2Probit(PercentAdjust(0, n))

#> [1] 2.424171