Viability constants.
Source: R/viabilityconstants.R

A dataset of the known species-specific seed viability constants estimated from the 'improved seed viability equation' of Ellis and Roberts (1980).

viabilityconstants

Format

A data frame with 7 columns:

Index

A unique index value, called by the vcindex argument in different functions.

Species

The species for which the constants were estimated.

Ke

The species-specific moisture constant KE

Cw

The species-specific moisture constant CW

Ch

The species-specific temperature constant CH

Cq

The species-specific temperature constant CQ

Reference

Source of the data.

Source

Primarily from Hong et al., 1996, updated with more species. The source for each record is mentioned in the Reference column.

Details

Various functions such as Sigma, StorageMC use the viability constants from this dataset for computation through the argument vcindex.

Note

Viability constants derived from storage experiment at a single temperature are highlighted by "*" in species column. In such cases, CH and CQ are the universal temperature coefficients (0.0329 and 0.000478 respectively).

References

Ellis R and Roberts E (1980). “Improved equations for the prediction of seed longevity.” Annals of Botany, 45(1), pp. 13--30.

Hong TD, Linington S and Ellis RH (1996). Seed Storage Behaviour: A Compendium, number 4 series Handbooks for Genebanks. International Plant Genetic Resources Institute (IPGRI), Rome, Italy.

Examples

# Soybean
viabilityconstants[grepl("glycine", x = viabilityconstants$Species,
                         ignore.case = TRUE),]

#> # A tibble: 2 x 7
#>   Index Species        Ke    Cw     Ch       Cq Reference                      
#>   <dbl> <chr>       <dbl> <dbl>  <dbl>    <dbl> <chr>                          
#> 1   59. Glycine max  7.52  4.09 0.0329 0.000478 Ellis et al., 1982; Dickie et ~
#> 2   60. Glycine max  7.29  4.00 0.0295 0.000491 Ellis, 1988                    

# Rice
viabilityconstants[grepl("oryza", x = viabilityconstants$Species,
                         ignore.case = TRUE),]

#> # A tibble: 8 x 7
#>   Index Species                        Ke    Cw     Ch       Cq Reference      
#>   <dbl> <chr>                       <dbl> <dbl>  <dbl>    <dbl> <chr>          
#> 1   85. Oryza glaberrima             9.41  5.04 0.0375 0.000471 Ellis et al., ~
#> 2   86. Oryza glaberrima*            6.87  5.51 0.0329 0.000478 Bam, Hong, Ell~
#> 3   87. Oryza sativa                 8.24  4.34 0.0307 0.000501 Ellis et al., ~
#> 4   88. Oryza sativa*                8.67  5.03 0.0329 0.000478 Ellis et al., ~
#> 5   89. Oryza sativa ssp. indica*    8.81  4.90 0.0329 0.000478 Ellis, Hong, a~
#> 6   90. Oryza sativa ssp. japonica*  8.42  4.90 0.0329 0.000478 Ellis et al., ~
#> 7   91. Oryza sativa ssp. japonica*  6.63  5.51 0.0329 0.000478 Bam et al., 20~
#> 8   92. Oryza sativa ssp. javanica*  8.74  4.90 0.0329 0.000478 Ellis et al., ~

# Wheat
viabilityconstants[grepl("triticum", x = viabilityconstants$Species,
                         ignore.case = TRUE),]

#> # A tibble: 2 x 7
#>   Index Species               Ke    Cw     Ch       Cq Reference         
#>   <dbl> <chr>              <dbl> <dbl>  <dbl>    <dbl> <chr>             
#> 1  136. Triticum aestivum   9.04  5.18 0.0351 0.000475 Ellis et al., 2007
#> 2  137. Triticum aestivum*  9.42  5.86 0.0329 0.000478 Ellis et al., 1990

# View dataset
View(viabilityconstants)