Viability constants

A dataset of the known species-specific seed viability constants estimated from the 'improved seed viability equation' of Ellis and Roberts (1980) .

viabilityconstants

Format

A data frame with 7 columns:

Index

A unique index value, called by the vcindex argument in different functions.

Species

The species for which the constants were estimated.

Ke

The species-specific moisture constant KE

Cw

The species-specific moisture constant CW

Ch

The species-specific temperature constant CH

Cq

The species-specific temperature constant CQ

Reference

Source of the data.

Source

Primarily from Hong et al. (1996) , updated with more species. The source for each record is mentioned in the Reference column.

Details

Various functions such as Sigma, StorageMC use the viability constants from this dataset for computation through the argument vcindex.

Note

Viability constants derived from storage experiment at a single temperature are highlighted by "*" in species column. In such cases, CH and CQ are the universal temperature coefficients (0.0329 and 0.000478 respectively).

References

Ali TS (2014). “Determination of seed viability constants in sorghum under various storage conditions.” Iranian Journal of Field Crop Science (Iranian Journal of Agricultural Sciences), 45(3), 377--387.

Alivand R, Tavakol Afshari R, Sharifzade F (2013). “Germination response and estimation of seed deterioration of Brassica napus under various storage conditions.” Iranian Journal of Field Crop Science, 44(1), 69--81.

Baladi S, Balouchi H (2016). “Evaluation of Lallemantia royleana seed longevity under varying conditions of temperature and moisture content.” Seed Science and Technology, 44(2), 320--326.

Balouchi H, Baladi S, Moradi A, Dehnavi MM (2017). “The influence of temperature and moisture content on seed longevity of two genotypes of Linum usitatissimum.” Seed Science and Technology, 45(1), 130--138.

Bam RK, Hong TD, Ellis RH, Kumaga FK, Asiedu EA (2008). “Storage behaviour of two contrasting upland rice genotypes.” Ghana Journal of Agricultural Science, 41(1).

Belletti P, Lanteri S, Lotito S (1991). “The influence of temperature and moisture on seed ageing in Iceland poppy (Papaver nudicaule L.).” Scientia Horticulturae, 48(1-2), 153--158.

Bonner FT (1994). “Predicting seed longevity for four forest tree species with orthodox seeds.” Proceedings of the International Seed Testing Association.

Chaves MMF, Usberti R (2004). “Controlled seed deterioration in Dalbergia nigra and Dimorphandra mollis, endangered Brazilian forest species.” Seed Science and Technology, 32(3), 813--823.

Crawford AD, Hay FR, Plummer JA, Probert RJ, Steadman KJ (2013). “One-step fitting of seed viability constants for two Australian plant species, Eucalyptus erythrocorys (Myrtaceae) and Xanthorrhoea preissii (Xanthorrhoeacea).” Australian Journal of Botany, 61(1), 1--10.

Daniel IO, Kruse M, Borner A (2011). “Comparative longevity and viability modeling of Solanum macrocarpon L. seeds.” Seed Science and Technology, 39(3), 680--685.

Daniel IO, Kruse M, Borner A (2013). “Controlled deterioration and predicting viability of okra seed in storage.” International Journal of Vegetable Science, 19(4), 324--333.

Daniel IO, Kruse M, Borner A (2012). “Predicting longevity of Celosia Argentea L. seeds during storage.” Acta Horticulturae, 953, 319--324.

Daniel IO, Ng NQ, Tayo TO, Togun AO (2003). “Storage of West African yam (Dioscorea spp.) seeds: modelling seed survival under controlled storage environments.” Seed Science and Technology, 31(1), 139--147.

Dehghan M, Sharif-Zadeh F (2015). “Viability model and effect of two drying procedures on seed longevity of Secale montanum seeds.” Biomedicine and Nursing, 1(1), 43--48.

Demir I, Kenanoglu BB, Hay F, Mavi K, Celikkol T (2011). “Determination of seed moisture constants (KE, CW) for the viability equation for watermelon, melon, and cucumber seeds.” Seed Science and Technology, 39(2), 527--532.

Demir I, Kenanoglu BB, Mavi K, Celikkol T, Hay F, Sariyildiz Z (2009). “Derivation of constants (KE, CW) for the viability equation for pepper seeds and the subsequent test of its applicability.” HortScience, 44(6), 1679--1682.

Dickie JB (1988). “Prospects for the long-term storage of apple seeds.” Veroffentlichungen der Landwirtschajthch-Chemischen Bundesanstalt Linz/Donau, 19, 47--63.

Dickie JB, Bowyer JT (1985). “Estimation of provisional seed viability constants for apple (Malus domestica borkh. cv. Greensleeves).” Annals of Botany, 56(2), 271--275.

Dickie JB, Ellis RH, Kraak HL, Ryder K, Tompsett PB (1990). “Temperature and seed storage longevity.” Annals of Botany, 65(2), 197--204.

Dickie JB, May K, Morris SVA, Titley SE (1991). “The effects of desiccation on seed survival in Acer platanoides L. and Acer pseudoplatanus L.” Seed Science Research, 1(03).

Dickie JB, Smith RD (1995). “Observations on the survival of seeds of Agathis spp. stored at low moisture contents and temperatures.” Seed Science Research, 5(01).

Ellis RH (1988). “The viability equation, seed viability nomographs, and practical advice on seed storage.” Seed Science and Technology, 16(1), 29--50.

Ellis RH, Hong TD (2007). “Quantitative response of the longevity of seed of twelve crops to temperature and moisture in hermetic storage.” Seed Science and Technology, 35(2), 432--444.

Ellis RH, Hong TD, Roberts EH (1989). “A comparison of the low-moisture-content limit to the logarithmic relation between seed moisture and longevity in twelve species.” Annals of Botany, 63(6), 601--611.

Ellis RH, Hong TD, Roberts EH (1986). “Logarithmic relationship between moisture content and longevity in sesame seeds.” Annals of Botany, 57(4), 499--503.

Ellis RH, Hong TD, Roberts EH (1988). “A low-moisture-content limit to logarithmic relations between seed moisture content and longevity.” Annals of Botany, 61(4), 405--408.

Ellis RH, Hong TD, Roberts EH (1992). “The low-moisture-content limit to the negative logarithmic relation between seed longevity and moisture content in three subspecies of rice.” Annals of Botany, 69(1), 53--58.

Ellis RH, Hong TD, Roberts EH, Tao K (1990). “Low moisture content limits to relations between seed longevity and moisture.” Annals of Botany, 65(5), 493--504.

Ellis RH, Osei-Bonsu K, Roberts EH (1982). “The influence of genotype, temperature and moisture on seed longevity in chickpea, cowpea and soya bean.” Annals of Botany, 50(1), 69--82.

Ellis RH, Roberts EH (1980). “Improved equations for the prediction of seed longevity.” Annals of Botany, 45(1), 13--30.

Ellis RH, Roberts EH (1980). “The influence of temperature and moisture on seed viability period in barley (Hordeum Distichum l.).” Annals of Botany, 45(1), 31--37.

Ellis RH, Roberts EH (1981). “The quantification of ageing and survival in orthodox seeds.” Seed Science and Technology.

Fantinatti JB, Usberti R (2007). “Seed viability constants for Eucalyptus grandis.” Pesquisa Agropecuaria Brasileira, 42(1), 111--117.

Ghaderi FF, Soltani A, Sadeghipour HR (2010). “Determination of seed viability constants in medicinal pumpkin (Cucurbita pepo L. subsp. pepo. convar. pepo var. styriaca Greb), borago (Borago officinalis L.) and black cumin (Nigella sativa L.).” Journal of Plant Production (Journal of Agricultural Sciences and Natural Resources), 17(3), 53--66.

Hay FR, Mead A, Manger K, Wilson FJ (2003). “One-step analysis of seed storage data and the longevity of Arabidopsis thaliana seeds.” Journal of Experimental Botany, 54(384), 993--1011.

Hong TD, Linington S, Ellis RH (1996). Seed Storage Behaviour: A Compendium, number 4 in Handbooks for Genebanks. International Plant Genetic Resources Institute (IPGRI), Rome, Italy.

Ignacio VL (2013). Germinacao e conservacao de sementes de Balfourodendron riedelianum (Engler) Engler. Ph.D. thesis, Universidade Estadual do Oeste do Parana.

Kebreab E, Murdoch AJ (1999). “Effect of temperature and humidity on the longevity of Orobanche seeds.” Weed Research, 39(3), 199--211.

Kraak HL, Vos J (1987). “Seed viability constants for lettuce.” Annals of Botany, 59(3), 343--349.

Kruse M, Ghiasi KG, Schmohl S (2005). “The seed viability equation for analysing seed storage behaviour.”

Kundu M (2008). “Prediction of viability of seeds of Pongamia pinnata (Karanj) under controlled conditions.” Seed Science and Technology, 36(2), 481--485.

Kuo WHJ (1991). “On the prediction of the storage longevity of muskmelon seeds.” Memoirs of the College of Agriculture, 31, 22--29.

Kuo WHJ, Shan ML, Tseng MT (1990). “Effects of temperature and seed moisture-content on the longevity of sorghum seeds.” Journal of the Agricultural Association of China, 32--41.

Lee MH, Hong SH, Na CS, Kim JG, Kim TW, Lee YH (2017). “Analysis of seed storage data and longevity for Agastache rugosa.” Korean Journal of Environmental Biology, 35, 207--214.

Muthoka PN, Hay FR, Dida MM, Nyabundi JO, Probert RJ (2009). “Moisture content and the longevity of seeds of six Euphorbia species in open storage.” Seed Science and Technology, 37(2), 383--397.

Pozitano M, Usberti R (2009). “Seed controlled deterioration of three interspecific elephant grass pearl millet hybrids.” Revista Brasileira de Zootecnia, 38, 428--434.

Reza EH (2014). “Estimation of seed viability constants for tall wheatgrass, cocksfoot, rye, and sheep fescue to inform gene banking decisions.” Iranian Journal of Plant Physiology, 4(4), 1145--1149.

Reza EH (2014). “Prediction of seed regeneration time of some medicinal plants by estimation of viability equation constants.” Iranian Journal of Field Crop Science, 45(3), 399--407.

Simoes FC, Usberti R, Paiva PDO (2008). “Controlled seed deterioration in Tagetes patula L. cultivars.” Seed Science and Technology, 36(3), 524--533.

Sinicio R, Lopes JF, Silva DJH, Mattedi AP (2009). “Longevity equation for tomato seeds.” Seed Science and Technology, 37(3), 667--675.

Tabatabaei SA (2014). “Determination of seed viability constants in sorghum under various storage conditions.” Iranian Journal of Field Crop Science, 45(3), 377--387.

Tompsett PB (1984). “The effect of moisture content and temperature on the storage life of Araucaria columnaris.” Seed Science and Technology, 12, 801--816.

Tompsett PB (1986). “The effect of temperature and moisture content on the longevity of seed of Ulmus carpinifolia and Terminalia brassii.” Annals of Botany, 57(6), 875--883.

Tompsett PB (1992). “A review of the literature on storage of dipterocarp seeds.” Seed Science and Technology, 20(2), 251--267.

Usberti R (2007). “Performance of tropical forage grass (Brachiaria brizantha) dormant seeds under controlled storage.” Seed Science and Technology, 35(2), 402--413.

Usberti R, Gomes RBR (1998). “Seed viability constants for groundnut.” Annals of Botany, 82(5), 691--694.

Usberti R, Roberts EH, Ellis RH (2006). “Prediction of cottonseed longevity.” Pesquisa Agropecuaria Brasileira, 41(9), 1435--1441.

Wilson DO, McDonald MB (1989). “A probit planes method for analyzing seed deterioration data.” Crop Science, 29(2), 471--476.

Zewdie M, Ellis RH (1991). “Response of tef and niger seed longevity to storage temperature and moisture.” Seed Science and Technology, 19(2), 319--329.

Examples

# Soybean
viabilityconstants[grepl("glycine", x = viabilityconstants$Species,
                         ignore.case = TRUE),]

#> # A tibble: 2 x 7
#>   Index Species        Ke    Cw     Ch      Cq Reference                        
#>   <dbl> <chr>       <dbl> <dbl>  <dbl>   <dbl> <chr>                            
#> 1    59 Glycine max  7.52  4.09 0.0329 4.78e-4 Ellis et al., 1982; Dickie et al~
#> 2    60 Glycine max  7.29  4.00 0.0295 4.91e-4 Ellis, 1988                      

# Rice
viabilityconstants[grepl("oryza", x = viabilityconstants$Species,
                         ignore.case = TRUE),]

#> # A tibble: 8 x 7
#>   Index Species               Ke    Cw     Ch      Cq Reference                 
#>   <dbl> <chr>              <dbl> <dbl>  <dbl>   <dbl> <chr>                     
#> 1    85 Oryza glaberrima    9.41  5.04 0.0375 4.71e-4 Ellis et al., 2007        
#> 2    86 Oryza glaberrima*   6.87  5.51 0.0329 4.78e-4 Bam, Hong, Ellis, Kumaga,~
#> 3    87 Oryza sativa        8.24  4.34 0.0307 5.01e-4 Ellis et al., 2007        
#> 4    88 Oryza sativa*       8.67  5.03 0.0329 4.78e-4 Ellis et al., 1989        
#> 5    89 Oryza sativa ssp.~  8.81  4.90 0.0329 4.78e-4 Ellis, Hong, and Roberts,~
#> 6    90 Oryza sativa ssp.~  8.42  4.90 0.0329 4.78e-4 Ellis et al., 1992        
#> 7    91 Oryza sativa ssp.~  6.63  5.51 0.0329 4.78e-4 Bam et al., 2008          
#> 8    92 Oryza sativa ssp.~  8.74  4.90 0.0329 4.78e-4 Ellis et al., 1992        

# Wheat
viabilityconstants[grepl("triticum", x = viabilityconstants$Species,
                         ignore.case = TRUE),]

#> # A tibble: 2 x 7
#>   Index Species               Ke    Cw     Ch       Cq Reference         
#>   <dbl> <chr>              <dbl> <dbl>  <dbl>    <dbl> <chr>             
#> 1   136 Triticum aestivum   9.04  5.18 0.0351 0.000475 Ellis et al., 2007
#> 2   137 Triticum aestivum*  9.42  5.86 0.0329 0.000478 Ellis et al., 1990

# View dataset
View(viabilityconstants)